Abstract existing feature rules components to child component
